+++ logprep/ng/abc/event_metadata.py
@@ -1,5 +1,4 @@
 # pylint: disable=too-few-public-methods
-# -> deactivate temporarily pylint issue here
 
 """abstract module for event related datatypes"""